A few warning fixes when actually building under 2.4.9-pre2 + some     reformating

+++ linux/drmP.h
@@ -78,7 +78,7 @@
 
 /* page_to_bus for earlier kernels, not optimal in all cases */
 #ifndef page_to_bus
-#define page_to_bus(page)	(virt_to_bus(page_address(page)))
+#define page_to_bus(page)	((unsigned int)(virt_to_bus(page_address(page))))
 #endif
 
 /* DRM template customization defaults
